NuSynthesis

Documentation for NuSynthesis.

NuSynthesis.first_until_tol

NuSynthesis.gen_data

NuSynthesis.get_statistics

NuSynthesis.nubar_conv

NuSynthesis.nubar_heuristic

NuSynthesis.run_experiments

NuSynthesis.first_until_tol — Method

ind = first_until_tol(Na, θ, tol, dstatistics)

Computes the first index for which the worst relative error is smaller than tol.

dstatistics is a directory containing statistics. The index is returned for the experiment concerning Na x Na matrices with nubar_conv run with interpolation parameter θ.

NuSynthesis.gen_data — Method

(data, dstatistics) = gen_data(maxN, Niter, Nexps, θs, optimizationfactory)

Computes the relative tolerance of $ar u$ for randomly generated matrices using nubar_heuristic.

Example

The results in the article were generated using the following bit of code.

using GLPK
maxN = 8                    # Runs experiments for square 2^1n x 2^n matrices, for n = 1:maxN.
Niter = 10000               # Runs nubar_heuristic with 10000 iterations
Nexps = 500                 # Generates 500 random examples per matrix size
θs = Vector(0.2:0.1:.9)     # Runs nubar_heuristic with θ = 0.2, 0.3, ..., 0.9

data, dstatistics = gen_data(maxN, Niter, Nexps, θs, GLPK.Optimizer)

NuSynthesis.get_statistics — Function

(ubound, means, stds) = get_statistics(experiment::Matrix{Float64}, tol=1e-8)

Computes the maximum, mean and standard deviation of the relative error of a batch of experiments.

experiment is a matrix where the first column contains the true values of $\bar \nu$ for each experiment. Column n + 1 contains the values of the $n$th iteration of the heuristic algorithm. The relative error of an estimate $\hat \nu$ of $\bar \nu$ is defined as $\hat \nu / \bar \nu + \text{tol}$.

NuSynthesis.nubar_conv — Method

(d, γ) = nubar_conv(A::Matrix{Float64}, optimizationfactory)

Computes the weigths and cost for a square matrix A using linear programming.

Example

using Hypatia
A = [1. 2.; 3. 4.]
(d, cost) = nubar_conv(A, Hypatia.Optimizer)

NuSynthesis.nubar_heuristic — Method

(ds, costs) = nubar_heuristic(A::Matrix{Float64}, N::Int64, θ::Float64)

Calculates and the balancing weights and the costs using the heuristic Algorithm.

Here, A is a square matrix, N is the number of iterations and 0 < θ < 1 is the interpolation parameter.

Example

A = [1. 2.; 3. 4.]
(ds, costs) = nubar_heuristic(A, 10, 0.5)

NuSynthesis.run_experiments — Method

experiments = run_experiments(Na::Int64, Niter::Int64, Nexps::Int64, θs::Vector{Float64}, optimizationfactory)

Computes the costs for randomly generated matrices for the heuristic algorithm and the LP.

Runs nubar_conv and nubar_heuristic for Nexps randomly generate Na x Na-matrices. nubar_conv is run for Niter iterations for each θ ∈ θs.